fix: typing errors
