+++ eodag/types/__init__.py
@@ -19,7 +19,9 @@
 
 from __future__ import annotations
 
+from collections import UserList
 from typing import (
+    TYPE_CHECKING,
     Annotated,
     Any,
     Literal,
@@ -31,7 +33,10 @@ from typing import (
     get_origin,
 )
 
+import boto3
 from annotated_types import Gt, Lt
+from boto3 import Session
+from botocore.handlers import disable_signing
 from pydantic import BaseModel, ConfigDict, Field, create_model
 from pydantic.annotated_handlers import GetJsonSchemaHandler
 from pydantic.fields import FieldInfo
@@ -40,6 +45,9 @@ from pydantic_core import CoreSchema, PydanticUndefined
 
 from eodag.utils import copy_deepcopy
 from eodag.utils.exceptions import ValidationError
+
+if TYPE_CHECKING:
+    from mypy_boto3_s3 import S3ServiceResource
 
 # Types mapping from JSON Schema and OpenAPI 3.1.0 specifications to Python
 # See https://spec.openapis.org/oas/v3.1.0#data-types
@@ -447,3 +455,91 @@ class S3SessionKwargs(TypedDict, total=False):
     aws_secret_access_key: Optional[str]
     aws_session_token: Optional[str]
     profile_name: Optional[str]
+
+
+class S3AuthContext:
+
+    """Class defining an S3 authentication context (resource, session and authentication type used)"""
+
+    def __init__(
+        self, s3_resource: S3ServiceResource, auth_type: str, s3_session: Session
+    ):
+        self.s3_resource: S3ServiceResource = s3_resource
+        self.auth_type: str = auth_type
+        self.s3_session: Session = s3_session
+
+    @staticmethod
+    def create_auth_context_unsigned(endpoint_url: str):
+        """Auth strategy using no-sign-request"""
+        s3_resource = boto3.resource(service_name="s3", endpoint_url=endpoint_url)
+        s3_resource.meta.client.meta.events.register(
+            "choose-signer.s3.*", disable_signing
+        )
+        return S3AuthContext(
+            s3_resource=s3_resource, auth_type="unsigned", s3_session=Session()
+        )
+
+    @staticmethod
+    def create_auth_context_auth_profile(endpoint_url: str, profile_name: str):
+        """Auth strategy using ``aws_profile`` from provided credentials"""
+        s3_session = Session(profile_name=profile_name)
+        s3_resource = s3_session.resource(
+            service_name="s3",
+            endpoint_url=endpoint_url,
+        )
+        return S3AuthContext(
+            s3_resource=s3_resource, auth_type="auth_profile", s3_session=s3_session
+        )
+
+    @staticmethod
+    def create_auth_context_auth_keys(endpoint_url: str, credentials: dict[str, str]):
+        """Auth strategy using ``aws_access_key_id``/``aws_secret_access_key`` from provided credentials"""
+        s3_session_kwargs: S3SessionKwargs = {
+            "aws_access_key_id": credentials["aws_access_key_id"],
+            "aws_secret_access_key": credentials["aws_secret_access_key"],
+        }
+        if credentials.get("aws_session_token"):
+            s3_session_kwargs["aws_session_token"] = credentials["aws_session_token"]
+        s3_session = Session(**s3_session_kwargs)
+        s3_resource = s3_session.resource(
+            service_name="s3",
+            endpoint_url=endpoint_url,
+        )
+        return S3AuthContext(
+            s3_resource=s3_resource, auth_type="auth_keys", s3_session=s3_session
+        )
+
+    @staticmethod
+    def create_auth_context_env(endpoint_url: str):
+        """Auth strategy using current environment"""
+
+        s3_session = Session()
+        s3_resource = s3_session.resource(service_name="s3", endpoint_url=endpoint_url)
+        return S3AuthContext(
+            s3_resource=s3_resource, auth_type="env", s3_session=s3_session
+        )
+
+
+class S3AuthContextPool(UserList):
+
+    """Instances of S3AuthContextPool contain a list of possible authentication contexts
+    based on the given credentials"""
+
+    def __init__(self, endpoint_url, credentials: dict[str, str], *args: Any):
+        super(S3AuthContextPool, self).__init__(*args)
+        self.data.append(S3AuthContext.create_auth_context_unsigned(endpoint_url))
+        if "aws_profile" in credentials:
+            self.data.append(
+                S3AuthContext.create_auth_context_auth_profile(
+                    endpoint_url, credentials["aws_profile"]
+                )
+            )
+        if (
+            "aws_access_key_id" in credentials
+            and "aws_secret_access_key" in credentials
+        ):
+            self.data.append(
+                S3AuthContext.create_auth_context_auth_keys(endpoint_url, credentials)
+            )
+        self.data.append(S3AuthContext.create_auth_context_env(endpoint_url))
+        self.used_method = ""
